refactor(public-rsvp): hoist canPublicRsvp to models + ensureHttps on confirmation links (Issue 496)
