+++ frontend/src/screens/events/EventDetailScreen.tsx
@@ -5,7 +5,13 @@ import { extractApiError, getApiStatus } from '@/api/apiErrors';
 import { useEvent } from '@/api/events';
 import { useAuthStore } from '@/auth/store';
 import type { Event } from '@/models/event';
-import { canManageEvent, EventStatus, EventType, EventVisibility } from '@/models/event';
+import {
+  canManageEvent,
+  canPublicRsvp,
+  EventStatus,
+  EventType,
+  EventVisibility,
+} from '@/models/event';
 import { ContentContainer, ContentError, ContentLoading } from '@/screens/public/ContentContainer';
 import { formatEventDateTime } from '@/utils/datetime';
 import { linkifyText } from '@/utils/linkifyText';
@@ -16,7 +22,7 @@ import { EventDetailKebabMenu } from './EventDetailKebabMenu';
 import { EventMemberSection } from './EventMemberSection';
 import { EventTagChips } from './EventTagChips';
 import { EventPollCard } from './poll/EventPollCard';
-import { canPublicRsvp, PublicRsvpSection } from './PublicRsvpSection';
+import { PublicRsvpSection } from './PublicRsvpSection';
 
 function photoSrc(url: string, updatedAt: string | null): string {
   if (!updatedAt) return url;
