+++ frontend/src/screens/events/PublicRsvpSection.tsx
@@ -1,20 +1,10 @@
 import { useState } from 'react';
 
 import type { PublicRsvpOut } from '@/api/publicRsvp';
-import { type Event, EventStatus, EventType, EventVisibility } from '@/models/event';
+import type { Event } from '@/models/event';
 
 import { PublicRsvpConfirmation } from './PublicRsvpConfirmation';
 import { PublicRsvpForm } from './PublicRsvpForm';
-
-export function canPublicRsvp(event: Event): boolean {
-  return (
-    event.eventType === EventType.Official &&
-    event.visibility === EventVisibility.Public &&
-    event.rsvpEnabled &&
-    event.status !== EventStatus.Cancelled &&
-    !event.isPast
-  );
-}
 
 interface Props {
   event: Event;
